feat(contracts): add uint precision for transaction value parameter

+++ contracts/contracts/ShadowGroup.sol
@@ -11,7 +11,7 @@ contract ShadowGroup {
 
     struct Transaction {
         address to;
-        uint value;
+        uint256 value;
         bytes data;
         bool executed;
         uint numConfirmations;
